backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
ecdsa==0.19.1
idna==3.11
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
python-decouple==3.8
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
whitenoise==6.11.0
